Backend Order & Order items entities

+++ Backend/Spring-boot-ecommerce/src/main/java/com/ecommerce/angularspring/Springbootecommerce/dao/ProductRepository.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import com.ecommerce.angularspring.Springbootecommerce.entity.Product;
 
 @CrossOrigin("http://localhost:4200")
-public interface ProductRepository extends JpaRepository<Product, Long>{
+public interface ProductRepository extends JpaRepository<Product, Integer>{
 	
 	Page<Product> findByCategoryId(@Param("id") Long id, Pageable pageable);
